Refactor MSSQL connectivity and param propagation tests
Split into seperate files the MSSQL connectivity and param propagation tests. Use the mock mssql service for param propagation tests, a much cleaner abstraction.
